+++ libs/qtdlib/messages/requests/qtdviewmessagesrequest.h
@@ -8,11 +8,11 @@ class QTdViewMessagesRequest : public QTdRequest
 {
     Q_OBJECT
     qint64 m_chatId;
-    QList<qint32> m_messageIds;
+    QList<qint64> m_messageIds;
 
 public:
     explicit QTdViewMessagesRequest(QObject *parent = nullptr);
-    void setMessageIds(const QList<qint32> &messageIds);
+    void setMessageIds(const QList<qint64> &messageIds);
     void setChatId(const qint64 &id);
     QJsonObject marshalJson() Q_DECL_FINAL;
 };
